Reorganize golang project structure
