+++ go.mod
@@ -1,9 +1,9 @@
-module translit
+module github.com/eevan78/translit
 
 go 1.22.1
 
 require (
+	github.com/cavaliergopher/grab/v3 v3.0.1
 	github.com/porfirion/trie v1.0.0
 	golang.org/x/net v0.24.0
-	github.com/cavaliergopher/grab/v3 v3.0.1
 )
