+++ test/testdata/rec_godine.txt
@@ -1,0 +1,13 @@
+Oksford University Press, poznata po izdanju Oksfordskog rečnika, proglasila je reč "riz" (rizz) za reč godine. Ova odluka ističe značajnost i popularnost ovog izraza koji nove generacije sve češće koriste.
+
+Generacija Z ovu reč koristi kako bi opisala nečiju veštinu privlačenja ili zavođenja druge osobe. 
+
+Pretpostavlja se da reč "riz" (rizz) potiče od reči "harizma" i da se može koristiti kao glagol, kao na primer "rizz up" što znači "privući, zavesti, flertovati sa nekim", saopštila je izdavačka kuća, preneo je AP. 
+
+Reč "riz" potiče iz engleskog žargona i označava šarm i privlačnost, najčešće je korišćena među mladima. 
+
+Na početku, postojalo je osam potencijalnih kandidata za reč godine, a nakon internet glasanja, četiri su se našle u užem izboru. Konačni izbor za reč godine doneli su stručnjaci za jezik, izveštava Bi-Bi-Si.
+
+Jutjuber i strimer na platformi X Kai Senat značajno je doprineo popularizaciji reči "riz", jer ju je koristio unutar svog prijateljskog kruga.
+
+Ova godina pokazuje da reči koje nastaju u internet zajednici sve više utiču na naš svakodnevni jezik, naglasila je izdavačka kuća Oksford University Press.
